+++ ChangeLog.md
@@ -1,5 +1,6 @@
 #### v.4.0.1 (TBD)
 * internally identify dicom servers by port/ipadress combination instead of only port (#699)
+* DicomDirectory.AddFile returns a reference to the newly added patient-, study-, series- and instance-Record.
 * Update Json DS validation regex (#643)
 * Add option to DicomFile.Open how to deal with large tags (#617)
 * DicomFileMetaInformation now uses pre-existing SourceAET where possible